requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
alembic
faster-whisper
python-multipart
soundfile
pytest
pytest-asyncio
httpx
python-dotenv
aiosqlite